custom-elements/reactions/AriaMixin-string-attributes-historical.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {
  assert_false("ariaRelevant" in Element.prototype);
});
</script>
